feat: User proxy 추가

+++ community-aggregate/src/main/java/com/sunic/community/aggregate/community/logic/CommunityLogic.java
@@ -8,9 +8,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.sunic.community.aggregate.community.store.CommunityStore;
 import com.sunic.community.aggregate.community.store.MemberStore;
+import com.sunic.community.aggregate.proxy.UserProxy;
 import com.sunic.community.spec.community.entity.Community;
 import com.sunic.community.spec.community.entity.Member;
 import com.sunic.community.spec.community.exception.MembershipException;
+import com.sunic.community.spec.common.exception.UnauthorizedException;
 import com.sunic.community.spec.community.facade.sdo.CommunityCdo;
 import com.sunic.community.spec.community.facade.sdo.CommunityRdo;
 import com.sunic.community.spec.community.facade.sdo.CommunityUdo;
@@ -26,15 +28,18 @@ public class CommunityLogic {
 
 	private final CommunityStore communityStore;
 	private final MemberStore memberStore;
+	private final UserProxy userProxy;
 
 	@Transactional
 	public CommunityRdo registerCommunity(CommunityCdo communityCdo) {
+		validateAdminUser(communityCdo.getRegistrant());
 		Community community = communityStore.save(Community.create(communityCdo));
 		return convertToCommunityRdo(community);
 	}
 
 	@Transactional
 	public CommunityRdo modifyCommunity(CommunityUdo modifySdo) {
+		validateAdminUser(modifySdo.getModifier());
 		Community community = communityStore.findById(modifySdo.getId());
 		community.modify(modifySdo);
 		Community updated = communityStore.update(community);
@@ -42,7 +47,8 @@ public class CommunityLogic {
 	}
 
 	@Transactional
-	public void deleteCommunity(Integer communityId) {
+	public void deleteCommunity(Integer communityId, Integer userId) {
+		validateAdminUser(userId);
 		communityStore.deleteById(communityId);
 	}
 
@@ -108,5 +114,14 @@ public class CommunityLogic {
 			.modifier(community.getModifier())
 			.allowSelfJoin(community.isAllowSelfJoin())
 			.build();
+	}
+
+	private void validateAdminUser(Integer userId) {
+		if (userId == null) {
+			throw new UnauthorizedException("User ID is required");
+		}
+		if (!userProxy.checkUserIsAdmin(userId)) {
+			throw new UnauthorizedException("Admin privileges required for this operation");
+		}
 	}
 }
